Move MPSC API under fanring::mpsc
Put channel, sender, receiver, errors, and capacity configuration under the
explicit mpsc module. Move the implementation into its own source module and
update docs, tests, and benchmarks for the breaking API change.
